pyproject.toml
[tool.poetry]
name = "msimg"
version = "0.0.1"
description = "🎨 一个魔塔社区文生图API调用库，支持多模型、多API、容错重试、图床上传等高级特性"
authors = ["Xiaoqiang <xiaoqiangclub@hotmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/xiaoqiangclub/msimg"
repository = "https://github.com/xiaoqiangclub/msimg"
documentation = "https://github.com/xiaoqiangclub/msimg"
keywords = ["modelscope", "text-to-image", "ai", "image-generation", "text2img"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

[tool.poetry.dependencies]
python = "^3.8.1"
requests = "^2.31.0"
Pillow = "^10.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
flake8 = "^6.1.0"
mypy = "^1.5.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/xiaoqiangclub/msimg/issues"